+++ python_modules/libraries/dagster-dg/dagster_dg/cli/env.py
@@ -1,8 +1,6 @@
-from collections.abc import Mapping
 from pathlib import Path
 
 import click
-from dagster_shared.plus.config import DagsterPlusCliConfig
 from rich.console import Console
 from rich.table import Table
 
@@ -11,8 +9,6 @@ from dagster_dg.config import normalize_cli_config
 from dagster_dg.context import DgContext
 from dagster_dg.env import ProjectEnvVars
 from dagster_dg.utils import DgClickCommand, DgClickGroup
-from dagster_dg.utils.plus.gql import LOCAL_SECRETS_FILE_QUERY
-from dagster_dg.utils.plus.gql_client import DagsterCloudGraphQLClient
 from dagster_dg.utils.telemetry import cli_telemetry_wrapper
 
 
@@ -46,70 +42,3 @@ def list_env_command(**global_options: object) -> None:
         table.add_row(key, value)
     console = Console()
     console.print(table)
-
-
-# ########################
-# ##### PULL
-# ########################
-
-
-def _get_local_secrets_for_locations(
-    client: DagsterCloudGraphQLClient, location_names: set[str]
-) -> Mapping[str, Mapping[str, str]]:
-    secrets_by_location = {location_name: {} for location_name in location_names}
-
-    result = client.execute(LOCAL_SECRETS_FILE_QUERY)
-    for secret in result["secretsOrError"]["secrets"]:
-        if not secret["localDeploymentScope"]:
-            continue
-        for location_name in location_names:
-            if len(secret["locationNames"]) == 0 or location_name in secret["locationNames"]:
-                secrets_by_location[location_name][secret["secretName"]] = secret["secretValue"]
-
-    return secrets_by_location
-
-
-@env_group.command(name="pull", cls=DgClickCommand)
-@dg_global_options
-def pull_env_command(**global_options: object) -> None:
-    """Pull environment variables from the current project."""
-    cli_config = normalize_cli_config(global_options, click.get_current_context())
-
-    dg_context = DgContext.for_workspace_or_project_environment(Path.cwd(), cli_config)
-    if not DagsterPlusCliConfig.exists():
-        raise click.UsageError(
-            "`dg env pull` requires authentication with Dagster Plus. Run `dg plus login` to authenticate."
-        )
-    config = DagsterPlusCliConfig.get()
-
-    project_ctxs = []
-    if dg_context.is_workspace:
-        project_ctxs = [
-            dg_context.for_project_environment(project_spec.path, cli_config)
-            for project_spec in dg_context.project_specs
-        ]
-    else:
-        project_ctxs = [dg_context]
-
-    gql_client = DagsterCloudGraphQLClient.from_config(config)
-    secrets_by_location = _get_local_secrets_for_locations(
-        gql_client, {project_ctx.project_name for project_ctx in project_ctxs}
-    )
-
-    projects_without_secrets = {project_ctx.project_name for project_ctx in project_ctxs}
-    for project_ctx in project_ctxs:
-        if secrets_by_location[project_ctx.project_name]:
-            env = ProjectEnvVars.empty(project_ctx).with_values(
-                secrets_by_location[project_ctx.project_name]
-            )
-            env.write()
-            projects_without_secrets.remove(project_ctx.project_name)
-
-    if dg_context.is_project:
-        click.echo("Environment variables saved to .env")
-    else:
-        click.echo("Environment variables saved to .env for projects in workspace")
-        if projects_without_secrets:
-            click.echo(
-                f"Environment variables not found for projects: {', '.join(projects_without_secrets)}"
-            )
