+++ python_modules/libraries/dagster-dg/dagster_dg/utils/plus/gql.py
@@ -10,7 +10,7 @@ query CliDeploymentsQuery {
 
 LOCAL_SECRETS_FILE_QUERY = """
 {
-	secretsOrError {
+	viewableLocalSecretsOrError {
     __typename
     ... on Secrets {
       secrets {
@@ -19,6 +19,13 @@ LOCAL_SECRETS_FILE_QUERY = """
         secretValue
         locationNames
       }
+    }
+    ... on UnauthorizedError {
+        message
+    }
+    ... on PythonError {
+        message
+        stack
     }
   }
 }
