+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/plus_tests/utils.py
@@ -14,9 +14,15 @@ def mock_gql_response(
         json_body = json.loads(request_body) if request_body else {}
         body_query_normalized = [line.strip() for line in json_body["query"].strip().split("\n")]
         query_normalized = [line.strip() for line in query.strip().split("\n")]
+        neq_index = next(
+            (i for i, (a, b) in enumerate(zip(body_query_normalized, query_normalized)) if a != b),
+            None,
+        )
         return (
             body_query_normalized == query_normalized,
-            f"'{body_query_normalized}'\n!=\n'{query_normalized}'",
+            f"\n'{body_query_normalized[neq_index]}'\n!=\n'{query_normalized[neq_index]}'\n"
+            if neq_index is not None
+            else "",
         )
 
     responses.add(
